requirements.txt
telethon
cryptography
python-dotenv
pyyaml
